+++ src/config/configmanager.cpp
@@ -177,7 +177,6 @@ bool ConfigManager::load() {
 	loadBoolConfig(L, EXPERT_PVP_CANWALKTHROUGHMAGICWALLS, "canWalkThroughMagicWalls", false);
 	loadBoolConfig(L, SPELL_NAME_INSTEAD_WORDS, "spellNameInsteadOfWords", false);
 	loadBoolConfig(L, LOG_PLAYERS_STATEMENTS, "logPlayersStatements", false);
-	loadBoolConfig(L, TOGGLE_TEST_MODE, "toggleTestMode", false);
 	loadBoolConfig(L, ROOK_SYSTEM, "toggleRookSystem", false);
 	loadBoolConfig(L, TOGGLE_ADD_ROOK_ITEMS, "toggleAddRookItems", false);