+++ src/creatures/players/player.cpp
@@ -166,6 +166,8 @@ std::string Player::getDescription(int32_t lookDistance) {
 
 		if (group->access) {
 			s << " You are " << group->name << '.';
+		} else if (hasFlag(PlayerFlags_t::IsGameTester)) {
+			s << " You are a " << group->name << '.';
 		} else if (vocation->getId() != VOCATION_NONE) {
 			s << " You are " << vocation->getVocDescription() << '.';
 		} else {
@@ -189,6 +191,8 @@ std::string Player::getDescription(int32_t lookDistance) {
 
 		if (group->access) {
 			s << " " << getSubjectVerb() << " " << group->name << '.';
+		} else if (hasFlag(PlayerFlags_t::IsGameTester)) {
+			s << " " << getSubjectVerb() << " a " << group->name << '.';
 		} else if (vocation->getId() != VOCATION_NONE) {
 			s << " " << getSubjectVerb() << " " << vocation->getVocDescription() << '.';
 		} else {
@@ -6184,7 +6188,7 @@ bool Player::lastHitIsPlayer(const std::shared_ptr<Creature> &lastHitCreature) {
 }
 
 void Player::changeHealth(int32_t healthChange, bool sendHealthChange /* = true*/) {
-	if (g_configManager().getBoolean(TOGGLE_TEST_MODE)) {
+	if (hasFlag(PlayerFlags_t::IsGameTester)) {
 		return;
 	}
 
@@ -6193,7 +6197,7 @@ void Player::changeHealth(int32_t healthChange, bool sendHealthChange /* = true*
 }
 
 void Player::changeMana(int32_t manaChange) {
-	if (g_configManager().getBoolean(TOGGLE_TEST_MODE)) {
+	if (hasFlag(PlayerFlags_t::IsGameTester)) {
 		return;
 	}
 
@@ -6205,7 +6209,7 @@ void Player::changeMana(int32_t manaChange) {
 }
 
 void Player::changeSoul(int32_t soulChange) {
-	if (g_configManager().getBoolean(TOGGLE_TEST_MODE)) {
+	if (hasFlag(PlayerFlags_t::IsGameTester)) {
 		return;
 	}
 
@@ -7325,9 +7329,16 @@ uint8_t Player::getRandomMountId() const {
 		}
 	}
 
-	const auto playerMountsSize = static_cast<int32_t>(playerMounts.size() - 1);
-	const auto randomIndex = uniform_random(0, std::max<int32_t>(0, playerMountsSize));
-	return playerMounts.at(randomIndex);
+	if (playerMounts.empty()) {
+		return 0;
+	}
+
+	const auto randomIndex = uniform_random(0, static_cast<int32_t>(playerMounts.size() - 1));
+	if (randomIndex >= 0 && static_cast<size_t>(randomIndex) < playerMounts.size()) {
+		return playerMounts[randomIndex];
+	}
+
+	return 0;
 }
 
 bool Player::toggleMount(bool mount) {